+++ src/reducers/NotifReducer.js
@@ -1,0 +1,19 @@
+
+const initialState = [{
+  id: "1",
+  message: "Coucou"
+},
+{
+  id: "2",
+  message:"Salut"
+}]
+
+//console.log(initialState)
+
+const NotifReducer = (prevState= initialState, action)=> {
+ 
+  return prevState
+}
+
+
+export default NotifReducer
