--- src/reducers/navbarTopNotifReducer.js
@@ -1,19 +1,0 @@
- //import   from  "../actions/userAction"
-
-
-const initialState = [{
-  id: "1",
-  message: "Coucou"
-},
-{
-  id: "2",
-  message:"Salut"
-}]
-
-const navbarTopNotifReducer = (prevState= initialState, action)=> {
- 
-  return prevState
-}
-console.log(this.prevState)
-
-export default navbarTopNotifReducer
