--- src/reducers/navbarTopNotifReducer.test.js
@@ -1,25 +1,0 @@
-import navbarTopNotifReducer from "./navbarTopNotifReducer"
-
-
-describe("navbarTopNotifReducer", () => {
-  it("should return the number of notifications",() => {
-    const initialState = [{
-      id: "1",
-      message: "Coucou"
-    },
-    {
-      id: "2",
-      message:"Salut"
-    }]
-
-    const anyAction = {
-      type: "ANY_ACTION"
-    }
-
-    expect(navbarTopNotifReducer (initialState, anyAction)).toEqual(initialState)
-  })
-
-    
-
-
-
